laboratorio testing obligatorio improvements
