+++ testing/testing-project/config/test/jest.js
@@ -1,6 +1,7 @@
 module.exports = {
   rootDir: '../../',
   verbose: true,
+  preset: 'ts-jest',
   restoreMocks: true,
   testEnvironment: 'jsdom',
   setupFilesAfterEnv: ['<rootDir>/config/test/setup-after.ts'],
