+++ testing/testing-project/package.json
@@ -11,7 +11,7 @@
     "analyze": "npm run clean && npm run type-check && webpack --config ./config/webpack/analyze.js",
     "type-check": "tsc --noEmit",
     "type-check:watch": "npm run type-check -- --watch",
-    "test": "jest -c ./config/test/jest.js --verbose",
+    "test": "jest -c config/test/jest.js",
     "test:watch": "npm run test -- --watchAll -i",
     "clean": "rimraf dist"
   },
@@ -56,19 +56,20 @@
     "@emotion/babel-plugin": "^11.10.5",
     "@pmmmwh/react-refresh-webpack-plugin": "^0.5.10",
     "@testing-library/dom": "^8.20.0",
-    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
-    "@types/jest": "^29.4.0",
+    "@types/jest": "^29.5.12",
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
     "babel-loader": "^9.1.2",
     "dotenv-webpack": "^8.0.1",
     "html-webpack-plugin": "^5.5.0",
-    "jest": "^29.4.1",
-    "jest-environment-jsdom": "^29.4.1",
+    "jest": "^29.7.0",
+    "jest-environment-jsdom": "^29.7.0",
     "npm-run-all": "^4.1.5",
     "react-refresh": "^0.14.0",
     "rimraf": "^4.1.2",
+    "ts-jest": "^29.1.2",
     "typescript": "^4.9.5",
     "webpack": "^5.75.0",
     "webpack-bundle-analyzer": "^4.7.0",
